Updates from code review: + Added ExportManager and ExportFormat in export package + Filename can be either string or closure + Filetype throws if not expected type + Show 'Export Starting' toast for larger exports to provide immediate user feedback
