+++ columns/DatesTimes.js
@@ -6,7 +6,7 @@
  */
 import {fileColFactory} from './Utils.js';
 import {dateRenderer, dateTimeRenderer, timeRenderer, compactDateRenderer} from '../format';
-import {exportFormats} from '@xh/hoist/utils/ExportUtils';
+import {ExportFormat} from '@xh/hoist/export';
 
 const colFactory = fileColFactory({
     cellStyle: {align: 'right'}
@@ -15,7 +15,7 @@ const colFactory = fileColFactory({
 export const dateCol = colFactory({
     headerName: 'Date',
     valueFormatter: dateRenderer(),
-    exportFormat: exportFormats.DATE_FMT,
+    exportFormat: ExportFormat.DATE_FMT,
     width: 120
 });
 
@@ -28,13 +28,13 @@ export const timeCol = colFactory({
 export const dateTimeCol = colFactory({
     headerName: 'Date',
     valueFormatter: dateTimeRenderer(),
-    exportFormat: exportFormats.DATETIME_FMT,
+    exportFormat: ExportFormat.DATETIME_FMT,
     width: 160
 });
 
 export const compactDateCol = colFactory({
     headerName: 'Date',
     valueFormatter: compactDateRenderer(),
-    exportFormat: exportFormats.DATE_FMT,
+    exportFormat: ExportFormat.DATE_FMT,
     width: 100
 });
