+++ columns/Utils.js
@@ -5,7 +5,7 @@
  * Copyright © 2018 Extremely Heavy Industries Inc.
  */
 import {Component} from 'react';
-import {exportFormats} from '@xh/hoist/utils/ExportUtils';
+import {ExportFormat} from '@xh/hoist/export';
 import {castArray, defaults, isNumber, omit, startCase} from 'lodash';
 
 // Configs specific to / added by Hoist as extensions to ag-Grid's column API.
@@ -70,7 +70,7 @@ export function fileColFactory(fileVals = {}) {
             ret.exportName = ret.exportName || ret.headerName || startCase(ret.field);
 
             // Use default exportFormat if not defined
-            ret.exportFormat = ret.exportFormat || exportFormats.DEFAULT;
+            ret.exportFormat = ret.exportFormat || ExportFormat.DEFAULT;
 
             // Install fn to produce definition w/o custom Hoist configs for use by ag-grid,
             // which will complain about unknown config keys.
